Adds an option to specify compression used for pack:deb
This enables passing options to override what dpkg-deb uses by default,
which is zstd on newer versions of Debian/Ubuntu and xz on older
versions. Debian 11 can't read zstd so when building on newer Debian
base for older, specifying may be necessary.

This should aid in the resolution of heroku/cli#2240.